separate the PlantService and Protocol from the NetworkService file
